[#4329] Fix Wsign-compare in iface_mgr_linux.cc

+++ src/lib/dhcp/iface_mgr_linux.cc
@@ -346,7 +346,7 @@ void Netlink::rtnl_process_reply(NetlinkMessages& info) {
         }
 
         nlmsghdr* header = static_cast<nlmsghdr*>(static_cast<void*>(buf));
-        while (NLMSG_OK(header, status)) {
+        while (NLMSG_OK(header, static_cast<unsigned int>(status))) {
 
             // Received a message not addressed to our process, or not
             // with a sequence number we are expecting.  Ignore, and
